# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: PulseBoard API
  version: 1.0.0
  description: |
    The same API the PulseBoard web app uses.

    **Auth.** Sign up with `POST /auth/register`, confirm the emailed code with
    `POST /auth/verify-email`, then call `POST /auth/login` to get a short-lived access token and send it as
    `Authorization: Bearer <token>`. The refresh token is set as an httpOnly cookie
    scoped to `/api/v1/auth`; call `POST /auth/refresh` to rotate it and get a new
    access token.

    **Rate limits.** Auth endpoints allow 10 requests per minute per IP. Everything else
    allows 300 per minute per user, and AI endpoints 20 per minute per user. Responses
    carry `RateLimit-Limit` and `RateLimit-Remaining`; a 429 includes `Retry-After`.

    **Errors** always look like `{ "error": { "message": "...", "code": "..." } }`.

    **Realtime.** Connect with Socket.io to the same host, passing `auth: { token }`.
    Emit `board:join` with a board id to receive `task:created`, `task:updated`,
    `task:deleted`, `comment:created`, `presence` and `cursor:move` events for it.
servers:
  - url: /api/v1
...
security:
  - bearer: []
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: Workspaces you belong to
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items: { $ref: '#/components/schemas/Workspace' }
    post:
      tags: [Workspaces]
      summary: Create a workspace (you become its owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201': { description: Created }
  /workspaces/{workspaceId}:
    parameters: [{ $ref: '#/components/parameters/workspaceId' }]
    patch:
      tags: [Workspaces]
      summary: Rename a workspace
      description: Requires **admin**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Workspaces]
      summary: Delete a workspace and everything in it
      description: Requires **owner**. Deletes every board, task, comment and attachment in the workspace.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Error' }
  /workspaces/{workspaceId}/members:
    parameters: [{ $ref: '#/components/parameters/workspaceId' }]
    get:
      tags: [Workspaces]
      summary: List members
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  members:
                    type: array
                    items: { $ref: '#/components/schemas/Member' }
    post:
      tags: [Workspaces]
      summary: Add an existing user by email
      description: Requires **admin**. Admins can add members; only owners can add admins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [admin, member], default: member }
      responses:
        '201': { description: Added }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /workspaces/{workspaceId}/members/{userId}:
    parameters:
      - { $ref: '#/components/parameters/workspaceId' }
      - { name: userId, in: path, required: true, schema: { type: string } }
    patch:
      tags: [Workspaces]
      summary: Change a member's role
      description: You can only change the role of someone ranked below you, to a role below your own.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, member] }
      responses:
        '200': { description: OK }
        '403': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Workspaces]
      summary: Remove a member, or leave the workspace
      responses:
        '204': { description: Removed }
        '403': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId: { name: workspaceId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message: { type: string }
            code: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        username: { type: string }
        email: { type: string }
        color: { type: string }
    Member:
      allOf:
        - { $ref: '#/components/schemas/User' }
        - type: object
          properties:
            role: { type: string, enum: [owner, admin, member] }
    Workspace:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        role: { type: string, enum: [owner, admin, member] }
        memberCount: { type: integer }
...